ad-tracking-system/consumer/requirements.txt
pika
psycopg2-binary
prometheus_client
